con preguntas done

+++ app/Http/Livewire/Encuestas.php
@@ -5,6 +5,7 @@ namespace App\Http\Livewire;
 use Livewire\Component;
 use App\Models\Encuesta;
 use App\Models\Pregunta;
+use Illuminate\Support\Collection;
 
 class Encuestas extends Component
 {
@@ -13,12 +14,15 @@ class Encuestas extends Component
     public $editar = false;
     public $encuestaEditable;
     public $texto_modal = "Crear Encuesta";
-    public $vertodo= false;
+    public $vertodo = false;
     public $preguntas_encuesta = [];
     public $todaslaspreguntas = [];
     public $preguntas_id = [];
     public $tipos = [1 => "Si/No", 2 => "Multiple Seleccion", 3 => "Completado"];
+    public $encuesta_ver;
+    protected $listeners = ["add" => "add"];
     public function mount() {
+        $this->preguntas_encuesta = new \Illuminate\Database\Eloquent\Collection;
     }
     public function render()
     {
@@ -27,27 +31,33 @@ class Encuestas extends Component
         } else {
             $this->encuestas = Encuesta::orderBy('id', 'desc')->where('habilitado', 1)->get();
         }
-        $this->todaslaspreguntas = Pregunta::orderBy('id', 'desc')->get();
+
         return view('livewire.encuestas');
     }
 
     public function crear() {
         $this->validate();
-        Encuesta::create([
+        $encuesta = Encuesta::create([
             'nombre' => $this->nombre,
             'descripcion' => $this->descripcion,
             'habilitado' => 1
         ]);
-        $this->preguntas_encuesta = [];
+        if (count($this->preguntas_id) > 0) {
+            $encuesta->preguntas()->attach(
+                $this->preguntas_id
+            );
+        }
+        $this->preguntas_encuesta = new \Illuminate\Database\Eloquent\Collection;
         $this->limpiar();
         $this->dispatchBrowserEvent('encuestas'); 
     }
 
     public function limpiar() {
-        $this->preguntas_encuesta = [];
+        $this->preguntas_encuesta = new \Illuminate\Database\Eloquent\Collection;
         $this->preguntas_id= [];
         $this->nombre = '';
         $this->descripcion = '';
+        $this->texto_modal = "Crear Encuesta";
     }
 
     public function cancelar() {
@@ -58,6 +68,7 @@ class Encuestas extends Component
     public function editar(Encuesta $encuesta) {
         $this->texto_modal = "Actualizar Encuesta";
         $this->editar = true;
+        $this->emit('preguntas', $encuesta->id);
         $this->encuestaEditable = $encuesta;
         $this->nombre = $encuesta->nombre;
         $this->descripcion = $encuesta->descripcion;
@@ -99,24 +110,71 @@ class Encuestas extends Component
     public function updateTaskOrder($lists) {
         $this->preguntas_id= [];
         $auxiliar = [];
+        $auxiliar_collection = new \Illuminate\Database\Eloquent\Collection;
         foreach($lists as $list) {
             $id = (int)$list["value"];
             $this->preguntas_id[] = $id;
-            /*foreach ($this->preguntas_encuesta as $pregunta) {
-                if ($id == $pregunta->id) {
-                    $auxiliar[] = $pregunta;
-                    break;
+            if (!$this->editar) {
+                foreach ($this->preguntas_encuesta as $pregunta) {
+                    if ($id == $pregunta->id) {
+                        $auxiliar_collection->add($pregunta);
+                        break;
+                    }
                 }
-            }*/
+            }
         }
-        $this->encuestaEditable->preguntas()->detach();
-        $this->encuestaEditable->preguntas()->attach(
-            $this->preguntas_id
-        );
-        $this->preguntas_encuesta = Encuesta::find($this->encuestaEditable->id)->preguntas;
+        if ($this->editar) {
+            $this->encuestaEditable->preguntas()->detach();
+            $this->encuestaEditable->preguntas()->attach(
+                $this->preguntas_id
+            );
+            $this->preguntas_encuesta = Encuesta::find($this->encuestaEditable->id)->preguntas;
+        } else {
+            $this->preguntas_encuesta = $auxiliar_collection;
+        }
+    }
+
+    public function add($id) {
+        if ($this->editar) {
+            $this->encuestaEditable->preguntas()->attach(
+                [$id]
+            );
+            $this->preguntas_encuesta = Encuesta::find($this->encuestaEditable->id)->preguntas;
+        } else {
+            $this->preguntas_encuesta->add(Pregunta::find($id));
+            $this->preguntas_id[] = $id;
+        }
     }
 
     public function activar() {
         $this->dispatchBrowserEvent('popover-activar');
+    }
+
+    public function quitar($id) {
+        if ($this->editar) {
+            $this->encuestaEditable->preguntas()->detach(
+                [$id]
+            );
+            $this->preguntas_encuesta = Encuesta::find($this->encuestaEditable->id)->preguntas;
+        } else {
+            $indice = -1;
+            foreach ($this->preguntas_encuesta as $key => $pregunta) {
+                if ($id == $pregunta->id) {
+                    $indice = $key;
+                    break;
+                }
+            }
+            if ($indice > -1) {
+                $this->preguntas_encuesta->pull($indice);
+                $this->preguntas_id = array_filter($this->preguntas_id, function($valor) use ($id) {
+                    return $valor != $id;
+                });
+            }
+        }
+    }
+
+    public function verEncuesta(Encuesta $encuesta) {
+        $this->texto_modal = "Ver Encuesta";
+        $this->encuesta_ver = $encuesta;
     }
 }
